Module hematite_server::packet

MC Protocol packets

Modules

handshake

login

play

status

Structs

BlockChangeRecord

ChunkMeta

Stat

Enums

Direction

NextState

Traits

PacketRead
A trait for decoding any of the packet types in one ID namespace.

PacketWrite
A trait for encoding the body of a single packet type.

Protocol
A trait used for data which can be encoded/decoded as is.