Trait hematite_server::packet::PacketRead

pub trait PacketRead: Sized {
    fn inner_decode(src: &mut Read) -> Result<Self>;

    fn read<R: Read>(src: &mut R) -> Result<Self> { ... }
}

A trait for decoding any of the packet types in one ID namespace.

Required Methods

fn inner_decode(src: &mut Read) -> Result<Self>

Provided Methods

fn read<R: Read>(src: &mut R) -> Result<Self>

Reads a new packet from a reader, including length.

TODO: add support for compression.