Module hematite_server::types

Reexports

pub use self::arr::Arr;

pub use self::chunk::{ChunkColumn};

pub use self::pos::BlockPos;

pub use self::slot::Slot;

pub use self::uuid::UuidString;

pub use self::varnum::Var;

Modules

consts
MC Protocol constants.

Structs

ChatJson

Chunk
Chunk is a group of 16x16x16 blocks.

EntityMetadata
Entity Metadata Format

EntitySelector
An entity selector used in commands, for example @p or @e[type=Creeper,c=2].

NbtBlob
An object in the Named Binary Tag (NBT) file format.

Enums

NbtError
Errors that may be encountered when constructing, parsing, or encoding NbtValue and NbtBlob objects.

NbtValue
A value which can be represented in the Named Binary Tag (NBT) file format.