Struct hematite_server::types::EntityMetadata

pub struct EntityMetadata {
    // some fields omitted
}

Entity Metadata Format

All entities must send at least one item of metadata, in most cases this will be the health item.

The entity metadata format is quirky dictionary format, where the key and the value's type are packed in a single byte.

Note that entity metadata is a totally distinct concept from block metadata.

Methods

impl EntityMetadata

fn new() -> EntityMetadata